+++ composeApp/src/commonMain/kotlin/com/altmann/choresmanager/di/KoinModule.kt
@@ -1,9 +1,8 @@
 package com.altmann.choresmanager.di
 
 import com.altmann.choresmanager.database.UserDao
-import com.altmann.choresmanager.local.DatabaseDriverFactory
 import com.altmann.choresmanager.local.LocalDatabase
-import com.altmann.choresmanager.ui.screens.calendar.CalendarViewModel
+import com.altmann.choresmanager.ui.screens.home.HomeViewModel
 import org.koin.core.KoinApplication
 import org.koin.core.context.startKoin
 import org.koin.core.module.Module
@@ -18,7 +17,7 @@ val sharedModule = module {
         UserDao(get())
     }
     single<LocalDatabase> { LocalDatabase(get()) }
-    viewModelOf(::CalendarViewModel)
+    viewModelOf(::HomeViewModel)
 
 }