+++ composeApp/src/commonMain/kotlin/com/altmann/choresmanager/models/chores/Chore.kt
@@ -18,10 +18,9 @@ open class Chore(
     val title: String,
     val description: String,
     var priority: Priority,
-    val finishedDate: LocalDate?,
     val deleted: Boolean? = false,
     val finished: Boolean? = false,
-    val color: Color? = Color.DarkGray
+    val color: Color = Color.DarkGray
 ) {
     open fun duration() : Int{
         return endTime.compareTo(startTime)