+++ composeApp/src/commonMain/kotlin/com/altmann/choresmanager/models/chores/college/CollegeChore.kt
@@ -20,10 +20,9 @@ class CollegeChore(
     title: String,
     description: String,
     priority: Priority,
-    finishedDate: LocalDate,
     deleted: Boolean? = false,
     finished: Boolean? = false,
-    color: Color? = Color.Blue,
+    color: Color = Color.Blue,
 
     val subject: String,
     val location: String?,
@@ -35,7 +34,7 @@ class CollegeChore(
 ) : Chore(
     choreId, startTime, endTime, daysOfWeek,
     startDate, endDate, choreException, title, description,
-    priority, finishedDate, deleted, finished, color
+    priority, deleted, finished, color
 ) {
 
     init {