+++ composeApp/src/commonMain/kotlin/com/altmann/choresmanager/ui/screens/components/fields/TimeTextField.kt
@@ -1,7 +1,5 @@
-package com.altmann.choresmanager.ui.screens.components
+package com.altmann.choresmanager.ui.screens.components.fields
 
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.text.input.rememberTextFieldState
 import androidx.compose.material3.Text
 import androidx.compose.material3.TextField
 import androidx.compose.runtime.Composable
@@ -9,9 +7,7 @@ import androidx.compose.runtime.MutableState
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.text.TextRange
 import androidx.compose.ui.text.input.TextFieldValue
-import androidx.compose.ui.unit.dp
 import com.altmann.choresmanager.utils.DateTimeParser
-import kotlinx.coroutines.selects.select
 import kotlinx.datetime.LocalTime
 
 @Composable